Refactor: 컨트롤러에 있는 엔티티 검증 코드를 서비스 레이어로 이동

+++ src/main/java/com/prgrms/vouchermanagement/customer/CustomerService.java
@@ -40,8 +40,12 @@ public class CustomerService {
         return customerRepository.findById(customerId);
     }
 
-    public void removeCustomer(Long customerId) {
+    public boolean removeCustomer(Long customerId) {
+        if (!isRegisteredCustomer(customerId)) {
+            return false;
+        }
         customerRepository.remove(customerId);
+        return true;
     }
 
     /**
